piecesapp/requirements.txt
Django>=3.1.4
djangorestframework>=3.12.2
numpy>=1.19.4